fix: category controller because of nested categories

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -45,7 +45,9 @@ class CategoryController extends Controller
      */
     public function show(string $id)
     {
-        //
+        $this->categoryRepository->find($id);
+        $categories = $this->categoryRepository->getChildCategories($id);
+        return view('admin.categories.show', compact('categories','id'));
     }
 
     /**
@@ -70,7 +72,13 @@ class CategoryController extends Controller
      */
     public function destroy(string $id)
     {
-        $this->categoryRepository->delete($id);
-        return redirect()->back()->with('add-success',__('success_message.category.delete.success'));
+        $category = $this->categoryRepository->find($id);
+        if ($this->categoryRepository->getChildCategories($category->id)->count() > 0) {
+            return redirect()->back()->with('delete-failed', __('failed_messages.category.delete.failed'));
+        }else{
+            $this->categoryRepository->delete($id);
+            return redirect()->back()->with('delete-success',__('success_message.category.delete.success'));
+        }
+
     }
 }
